modificato l'AudioMonitor utilizzando un AudioRecord invece di un MediaRecorder

+++ app/src/main/java/com/example/mobile/monitors/IMonitor.kt
@@ -1,7 +1,8 @@
 package com.example.mobile.monitors
 
 interface IMonitor {
+    //TODO: maybe also add pause/resume functionality   
     fun startMonitoring()
     fun stopMonitoring()
-    fun readValue(): Int
+    fun readValue(): Double
 }
